requirements.txt
flask_bootstrap
flask_sqlalchemy
flask_wtf
requests~=2.28.1
Flask~=2.2.2
WTForms~=3.0.1
python-dotenv
